Support more general iter-real-structures
Order of iter vs real does not matter in
the filesystem, you only need to provide
a regexp that for each path is able to determine
the paths iteration string and realization integer.

Some stubs for support of a 'batch' level next
to the 'iteration' level. In the EnsembleSet
context iter and batch are to be mixed